# Conflicts:
#	javaBench/src/Controller/ScoreTabController.java
#	javaBench/src/Controller/TestTabController.java
#	javaBench/src/GUI/TabRoot.fxml
#	javaBench/src/GUI/TestTab.fxml
#	javaBench/src/Test/CPU/CompressTest.java
#	javaBench/src/Test/CPU/DataEncryptior.java
#	javaBench/src/Test/GPU/Window.java
#	javaBench/src/Test/HardDrive/MeasureIOPerformance.java

+++ javaBench/src/Helper/Timer.java
@@ -7,7 +7,7 @@ public class Timer {
 		play(); 
 	}
 	
-	public double check() { 
+	public long check() {
 		return (System.nanoTime()-start+spent); 
 	}
